# specs/001-enterprise-log-collector/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Log Collection API
  description: |
    RESTful API for enterprise log collection, transaction analysis, and alerting system.

    **Authentication**: OAuth2 Bearer tokens issued by Keycloak
    **Roles**: ADMIN, OPERATOR, VIEWER
  version: 1.0.0
  contact:
    name: API Support
    email: support@logcollector.com
servers:
  - url: https://api.logcollector.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local Development
security:
  - oauth2: []
...
paths:
...
  /transactions:
    get:
      summary: List transactions
      description: Retrieve paginated list of transactions with optional filtering
      tags: [Transactions]
      security:
        - oauth2: [read:transactions]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            default: "startTime,desc"
            example: "startTime,desc"
        - name: status
          in: query
          schema:
            type: string
            enum: [SUCCESS, FAILED, TIMEOUT, IN_PROGRESS]
        - name: startTimeFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: startTimeTo
          in: query
          schema:
            type: string
            format: date-time
        - name: sourceSystem
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTransactionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /transactions/{transactionId}:
    get:
      summary: Get transaction details
      description: Retrieve complete transaction information including associated log entries
      tags: [Transactions]
      security:
        - oauth2: [read:transactions]
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z0-9]+$'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /servers:
    get:
      summary: List server configurations
      description: Retrieve all configured log collection servers
      tags: [Servers]
      security:
        - oauth2: [read:servers]
      responses:
        '200':
          description: Server list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServerConfiguration'

    post:
      summary: Add server configuration
      description: Configure new server for log collection (ADMIN only)
      tags: [Servers]
      security:
        - oauth2: [admin:servers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServerRequest'
      responses:
        '201':
          description: Server configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerConfiguration'
        '400':
          $ref: '#/components/responses/BadRequest'
  /servers/{serverId}:
    put:
      summary: Update server configuration
      description: Modify server settings (ADMIN only)
      tags: [Servers]
      security:
        - oauth2: [admin:servers]
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
      responses:
        '200':
          description: Server updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerConfiguration'

    delete:
      summary: Remove server configuration
      description: Delete server from collection (ADMIN only)
      tags: [Servers]
      security:
        - oauth2: [admin:servers]
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Server removed

  # Archive Endpoints
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.example.com/realms/log-monitor/protocol/openid-connect/auth
          tokenUrl: https://keycloak.example.com/realms/log-monitor/protocol/openid-connect/token
          scopes:
            read:transactions: Read transaction data
            read:logs: Search and read logs
            read:alerts: View alerts
            write:alerts: Acknowledge alerts
            admin:alerts: Manage alert rules
            read:servers: View server configurations
            admin:servers: Manage server configurations
            read:archive: View archived data
            admin:archive: Trigger archival operations
            read:metrics: View system metrics
  schemas:
...
    PagedTransactionResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TransactionSummary'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
    TransactionSummary:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: [SUCCESS, FAILED, TIMEOUT, IN_PROGRESS]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          format: int64
          nullable: true
        sourceSystem:
          type: string
          nullable: true
        targetSystem:
          type: string
          nullable: true
    TransactionDetail:
      allOf:
        - $ref: '#/components/schemas/TransactionSummary'
        - type: object
          properties:
            errorMessage:
              type: string
              nullable: true
            retryCount:
              type: integer
            logEntries:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        serverName:
          type: string
        logPath:
          type: string
        logType:
          type: string
        content:
          type: string
        originalTimestamp:
          type: string
          format: date-time
        transactionId:
          type: string
          nullable: true
        highlight:
          type: string
          description: Highlighted search match
          nullable: true

    # Alert Schemas
...
    ServerConfiguration:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hostname:
          type: string
        port:
          type: integer
        username:
          type: string
        logFilePaths:
          type: array
          items:
            type: string
        collectionInterval:
          type: integer
        batchSize:
          type: integer
        status:
          type: string
          enum: [CONNECTED, DISCONNECTED, ERROR]
        lastConnectedAt:
          type: string
          format: date-time
          nullable: true
        lastCollectionAt:
          type: string
          format: date-time
          nullable: true
    CreateServerRequest:
      type: object
      required: [hostname, username, password, logFilePaths]
      properties:
        hostname:
          type: string
        port:
          type: integer
          default: 22
        username:
          type: string
        password:
          type: string
          format: password
        logFilePaths:
          type: array
          items:
            type: string
        collectionInterval:
          type: integer
          default: 30000
        batchSize:
          type: integer
          default: 500
    UpdateServerRequest:
      type: object
      properties:
        password:
          type: string
          format: password
        logFilePaths:
          type: array
          items:
            type: string
        collectionInterval:
          type: integer
        batchSize:
          type: integer

    # Archive Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'